requirements.txt
Django==1.7
South==1.0
argparse==1.2.1
djangorestframework==2.4.3
wsgiref==0.1.2
